Creates a ValidationResourceTest which is extended by CardExpiryValidationResourceTest, FirstSixCardDigitsValidationResourceTest, LastFourDigitsValidationResourceTest
